+++ CHANGELOG.md
@@ -56,5 +56,8 @@ __1.14.2024__
 __1.15.2024__
 
 - Partially implement the content scroller
+- Partially implement scroller visibility handling
 
 [Demo](https://drive.google.com/file/d/1qAY5l1wjrzedgmNRRUEOUVDsIGYDzlZR/view?usp=sharing)
+
+[Demo - It's a bit wonky still](https://drive.google.com/file/d/1qDUQdquDw5h_olay3i5-baa3D9lv21_g/view?usp=sharing)